should not emit an _assert helper if there are no asserts
