+++ lib/index.js
@@ -15,6 +15,7 @@ exports.default = function (_ref) {
 
   var tcombExpression = null;
   var assertHelperName = null;
+  var hasAsserts = false;
 
   var assertHelper = expression('\n    function assert(x, type, name) {\n      type = type || tcomb.Any;\n      if (tcomb.isType(type) && type.meta.kind !== \'struct\') {\n        type(x, [name + \': \' + tcomb.getTypeName(type)]);\n      } else if (!(x instanceof type)) {\n        tcomb.fail(\'Invalid value \' + tcomb.stringify(x) + \' supplied to \' + name + \' (expected a \' + tcomb.getTypeName(type) + \')\');\n      }\n      return x;\n    }\n  ');
 
@@ -330,6 +331,7 @@ exports.default = function (_ref) {
     var typeAnnotation = _ref4.typeAnnotation;
     var argumentName = _ref4.argumentName;
 
+    hasAsserts = true;
     var type = getType({ annotation: typeAnnotation, typeParameters: typeParameters });
     if (optional) {
       type = getMaybeCombinator(type);
@@ -497,6 +499,7 @@ exports.default = function (_ref) {
           // Ensure we reset the import between each file so that our guard
           // of the import works correctly.
           tcombExpression = null;
+          hasAsserts = false;
           assertHelperName = path.scope.generateUidIdentifier('assert');
         },
         exit: function exit(path, state) {
@@ -504,10 +507,12 @@ exports.default = function (_ref) {
             return;
           }
           ensureTcombExpression();
-          path.node.body.push(assertHelper({
-            assert: assertHelperName,
-            tcomb: tcombExpression
-          }));
+          if (hasAsserts) {
+            path.node.body.push(assertHelper({
+              assert: assertHelperName,
+              tcomb: tcombExpression
+            }));
+          }
         }
       },
