--- test/fixtures/assert/actual.js
@@ -1,5 +1,0 @@
-import t from 'tcomb';
-
-function _assert(){}
-
-function foo(x: string) {}
